Split JmsMessages and Settings files
